moved drink selector screen code into its own file and also only create new drink selector instance if one doesn't already exist
