+++ drink_selector.kv
@@ -2,7 +2,7 @@
 #:import kivy kivy
 #:import win kivy.core.window
 <DrinkSelectorScreen>:
-    on_pre_enter: self.add_drink_selector()
+    on_pre_enter: self.setup_drink_selector()
 
 <DrinkSelector>:
     cols: 2
